Addition of login logic and flow
When a new user signs up, they are redirected to the login page.

Here, they can proceed to the dashboard if their login credentials are verified.

If their credentials do not match up, an error message is displayed to the user.

A user object also gets added to state when login is successful for use by the app.